==> code/1_taxonomic_assignment_blastn.Rmd <==
---
title: "Taxonomic Assignment of NBS 2023 Mifish Amplicons"
author: "Kimberly Ledger"
date: "2025-01-29"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r load-libraries}
library(tidyverse)
library(tidyr)
library(stringr)
rename <- dplyr::rename
```

## load output from taxonkit 
- this is the fasta seqs from the 20250129 run on dadasnake of just the NBS23 samples, and blastn using the ncbi nt database with 96% seq identity and 98% query coverage   

```{r load-output-from-taxonkit}
taxonomy <-read.delim("/home/kimberly.ledger/NBS23_mifish_mb/data/blastn_taxlineage.txt", header = FALSE, na.strings=c(""," ","NA"))
head(taxonomy)
```


## clean up the header a bit
```{r}
# use the full taxonomy rather than the seq id to collapse identical entries
tax_df <- taxonomy %>%
  filter(V4 > 100) %>% # make sure all retained matches are >100 bp
  select(-V2, -V5, -V6, -V7, -V14, -V8, -V9, -V10, -V11, -V12, -V13, -V15, -V16) %>%  #remove unnecessary columns
  group_by(V1, V17) %>% # group by the sequence key and the full taxonomy to reduce duplicate entries
  unique() %>% # doing that reduced the number of entries
  rename(qseqid=V1, perc_id=V3, length=V4, taxonomy=V17) %>% #rename headers
  filter(perc_id > 96) %>% 
  filter(!is.na(taxonomy))
```

## calculate the number of ASV's with taxonomic assignment
```{r how-many-total-ASVs-w-taxonomy}
tax_df %>% 
  ungroup() %>%
  select(qseqid) %>%
  unique() %>%
  nrow
```

## reformat table 
```{r}
# formatting the taxonomy variables
taxon_df <- tax_df %>%
  filter(str_detect(taxonomy, ";")) %>%
  separate(taxonomy, into=c("kingdom", "phylum", "class", "order", "family", "genus", "species"), sep = ";") %>% 
  filter(!grepl("sp\\.", species, ignore.case = TRUE)) %>%
  filter(!grepl("cf\\.", species, ignore.case = TRUE)) %>%
  filter(!grepl("aff\\.", species, ignore.case = TRUE)) %>%
  filter(!grepl(" x ", species, ignore.case = TRUE)) %>%         ### remove hybrids and unknown species assignments
  filter(!grepl("\\d", species))
```

## what non-fish are here? 
```{r}
not_Actinopteri <- taxon_df %>%
  filter(class != "Actinopteri") 

unique(not_Actinopteri$species)
```

this particular project doesn't look at non-fish so will remove them all here
```{r}
taxon_df <- taxon_df %>%
  filter(class == "Actinopteri")
```


## now look at the fish and figure out what fish taxa are in/out of our range
```{r}
to_check_range <- taxon_df %>%
  ungroup() %>%
  select(species, genus, family, order, class, phylum) %>%
  unique()
```

## check ranges for species using rfishbase 
```{r}
#library(remotes)
#remotes::install_github("ropensci/rfishbase")
library(rfishbase)

#one-time download of all fishbase tables... this takes a bit 
#fb_import()

#first, validate species names using rfishbase synonyms
spp_df <- synonyms(to_check_range$species)

syn <- spp_df %>% 
  filter(Status == "synonym")

#rename any synonyms with the validated name  
to_check_range_validated <- to_check_range %>%
  rename(synonym = species) %>%
  left_join(syn, by = "synonym") %>%
  mutate(synonym = ifelse(is.na(Species), synonym, Species)) %>%
  select(synonym:phylum) %>%
  rename(species = synonym)

#get distribution info 
spp_distribution <- distribution(to_check_range_validated$species) %>%
  select(Species, FAO) %>%
  unique()

#add column to designate if we will consider a species as "in range"- for this study, this will be NE Pacific 
spp_distribution <- spp_distribution %>%
  mutate(in_range = ifelse(is.na(FAO), NA, "no"),
         in_range = ifelse(FAO == "Pacific, Northeast", "yes", in_range)) #,
         #in_range = ifelse(FAO == "America, North - Inland waters", "yes", in_range))

#keep just a list of spp names and yes/no/NA for "in range"  - this way we can keep track of what spp didn't have any reference information in fishbase to determine range 
spp_range <- spp_distribution %>%
  select(Species, in_range) %>%
  unique()

#how many entries do not have range info? 
range_na <- spp_range %>%
  filter(is.na(in_range))
```

When a valid name was not found, the presence of a species in the study area was checked using the GBIF database (https://www.gbif.org/).

## some species do not have range info - manually determine if these species should be considered in range
```{r}
spp_range <- spp_range %>%
  mutate(in_range = ifelse(Species == "natural gynogenetic Carassius auratus red var.", "no", in_range)) %>%
  mutate(in_range = ifelse(Species == "Ammodytes japonicus", "no", in_range)) %>%
  mutate(in_range = ifelse(Species == "Bothrocara zestum", "no", in_range))

in_range <- spp_range %>%
  filter(in_range == "yes") 
```

## keep only the "in range" fish species in the taxon dataframe 
```{r}
taxon_df_reduced <- taxon_df %>%
  filter(species %in% in_range$Species)
```

## figure out what ASVs are unambiguously a single species
```{r}
single_spp_seqs <- taxon_df_reduced %>% 
  group_by(qseqid) %>%
  add_tally(name = "n_taxa") %>%
  filter(n_taxa == 1)
```

## single species with at least 98% similarity 
```{r}
single_spp_seqs_98 <- single_spp_seqs %>%
  filter(perc_id >=98)
```

## remove the single-species seqs from the dataframe and then rank the hits by % identity for the remaining seqs
```{r}
seq_id_diff <- taxon_df_reduced %>%
  anti_join(., single_spp_seqs) %>%  ## comment out this line if there are no single species ASVs
  select(-length) %>%
  group_by(qseqid, species, genus, family, order, class, phylum, kingdom) %>%
    mutate(seq_percID = max(perc_id)) %>%
    group_by(qseqid, species, genus, family, order, class, phylum, kingdom, seq_percID) %>%
  summarise(max(seq_percID)) %>% # take just the top hit for each taxon (for each sequence)
  select(-`max(seq_percID)`) %>%
  ungroup() %>%
  group_by(qseqid) %>%
      #mutate(id_rank = rank(desc(seq_percID), ties.method = "min")) %>% # rank the taxonomic hits per sequence by % id
       mutate(top_perc = max(seq_percID)) %>% # designate the highest % id for the best taxonomic hit in each sequence (in some, but not all cases, this is 100%)   %>%
      mutate(diff = top_perc - seq_percID) %>% # calculate the difference between the % identity of the top hit and each subsequent taxonomic hit
        arrange(diff)
```

Now you have the single best entry for each species for each sequence ranked and with the difference between the first and second ranked entries calculated.
** if there is a sequence that is >98% match, then keep just the sequences within 0.5%  
** if there is NOT a >98% match, keep the top 1% 

## ASVs with >98% match or <98% match
```{r}
match_98 <- seq_id_diff %>%
  filter(top_perc >= 98) %>%
  select(qseqid) %>%
  unique()

match_lessthan98 <- seq_id_diff %>%
  filter(top_perc < 98) %>%
  select(qseqid) %>%
  unique()
```


## Figure out which differences are outside the threshold and eliminate those first
```{r}
# filter out any taxa that are >1% less matching identity than the top taxonomic hit for a given sequence
match_98_to_remove_low_perc_hits <- seq_id_diff %>%
  ungroup() %>%
  filter(qseqid %in% match_98$qseqid) %>%
  group_by(qseqid) %>%
  filter(diff > 0.5)

match_lessthan98_to_remove_low_perc_hits <- seq_id_diff %>%
  ungroup() %>%
  filter(qseqid %in% match_lessthan98$qseqid) %>%
  group_by(qseqid) %>%
  filter(diff > 1)

to_remove_low_perc_hits <- match_98_to_remove_low_perc_hits %>%
  bind_rows(match_lessthan98_to_remove_low_perc_hits)

keepers <- seq_id_diff %>%
  anti_join(to_remove_low_perc_hits)
```

## so now determine whether ASVs should be assigned to genus, family, order, etc. 
```{r}
singletons <- keepers %>%
  select(qseqid) %>%
  tally() %>%
  filter(n == 1)

# these are the seqs that now have only a single match
singleton_df <- singletons %>%
  left_join(keepers) %>%
  select(-n) %>%
  bind_rows(single_spp_seqs_98) %>% # combine the single spp data
  mutate(taxonomic_level = "species") %>%
  mutate(taxon = species)

## Genus-level matches
# remove the singletons from the bigger df 
single_genus <- keepers %>%
  anti_join(singleton_df)%>% # at best, these should be genus-level matches
  group_by(qseqid, genus) %>%
  tally() %>%
  ungroup() %>%
  group_by(qseqid) %>%
  tally() %>%
  filter(n == 1) %>% # seqs that match a single genus
  select(-n) %>%
  left_join(., keepers) %>%
  mutate(taxonomic_level = "genus") %>%
  mutate(taxon = genus)
  

## Family-level matches
single_family <- keepers %>%
  anti_join(singleton_df)%>%
  anti_join(single_genus) %>%
  group_by(qseqid, family) %>%
  tally() %>%
  ungroup() %>%
  group_by(qseqid) %>%
  tally() %>%
  filter(n == 1) %>% # seqs that match a single family
  select(-n) %>%
  left_join(., keepers) %>%
  mutate(taxonomic_level = "family") %>%
  mutate(taxon = family)


## Order-level matches
single_order <- keepers %>%
  anti_join(singleton_df)%>%
  anti_join(single_genus) %>%
  anti_join(single_family) %>%
  group_by(qseqid, order) %>%
  tally() %>%
  ungroup() %>%
  group_by(qseqid) %>%
  tally() %>%
  filter(n == 1) %>% # seqs that match a single order
  select(-n) %>%
  left_join(., keepers) %>%
  mutate(taxonomic_level = "order") %>%
  mutate(taxon = order)


## Class-level matches
single_class <- keepers %>%
  anti_join(singleton_df)%>%
  anti_join(single_genus) %>%
  anti_join(single_family) %>%
  anti_join(single_order) %>%
  group_by(qseqid, class) %>%
  tally() %>%
  ungroup() %>%
  group_by(qseqid) %>%
  tally() %>% 
  filter(n == 1) %>% # seqs that match a single class
  select(-n) %>%
  left_join(., keepers) %>%
  mutate(taxonomic_level = "class") %>%
  mutate(taxon = class)


## Phylum-level matches
single_phylum <- keepers %>%
  anti_join(singleton_df)%>%
  anti_join(single_genus) %>%
  anti_join(single_family) %>%
  anti_join(single_order) %>%
  anti_join(single_class) %>%
  group_by(qseqid, phylum) %>%
  tally() %>%
  ungroup() %>%
  group_by(qseqid) %>%
  tally() %>% 
  filter(n == 1) %>% # seqs that match a single phylum
  select(-n) %>%
  left_join(., keepers) %>%
  mutate(taxonomic_level = "phylum") %>%
  mutate(taxon = phylum)

# Kingdom level matches
single_kingdom <- keepers %>%
  anti_join(singleton_df)%>%
  anti_join(single_genus) %>%
  anti_join(single_family) %>%
  anti_join(single_order) %>%
  anti_join(single_class) %>%
  group_by(qseqid, kingdom) %>%
  tally() %>%
  ungroup() %>%
  group_by(qseqid) %>%
  tally() %>% 
  filter(n == 1) %>% # seqs that match a single kingdom
  select(-n) %>%
  left_join(., keepers) %>%
  mutate(taxonomic_level = "kingdom") %>%
  mutate(taxon = kingdom)
```

## modify the singleton_df to include the right variable headers
```{r}
single_spp <- singleton_df %>%
  select(-perc_id, -length, -n_taxa) %>%
  mutate(taxonomic_level = "species") %>%
  mutate(taxon = species)
```

## recombine the full data set now that the appropriate level of taxonomy has been determined
```{r}
tax_df_2 <- bind_rows(single_kingdom, single_phylum, single_class, single_order, single_family, single_genus, single_spp)
```

## Create output taxonomy data frames
```{r}
# and then collapse that down to just a single taxon per ASV
# can i collapse while keeping the taxonomy for the taxon....? 
tax_df_3 <- tax_df_2 %>%
  mutate(species = ifelse(taxonomic_level == "family", NA, species),
         genus = ifelse(taxonomic_level == "family", NA, genus),
         species = ifelse(taxonomic_level == "genus", NA, species)) %>%
  select(qseqid, taxon, taxonomic_level, species, genus, family, order, class, phylum, kingdom) %>%
  unique()
```

```{r}
write.csv(tax_df_2, "/home/kimberly.ledger/NBS23_mifish_mb/outputs/taxonomy_20250129_uncollapsed.csv")
write.csv(tax_df_3, "/home/kimberly.ledger/NBS23_mifish_mb//outputs/taxonomy_20250129_collapsed.csv")
```
